+++ addon/plugins/variable-plugin/variables/text.ts
@@ -76,7 +76,7 @@ export const emberNodeConfig: EmberNodeConfig = {
   name: 'text-variable',
   componentPath: 'variable-plugin/variable/nodeview',
   inline: true,
-  group: 'inline',
+  group: 'inline variable',
   content: 'inline*',
   atom: true,
   recreateUri: true,